\name{TwoPhaseInd-package}
\alias{TwoPhaseInd-package}
\alias{TwoPhaseInd}
\docType{package}
\title{
  Semiparametric Estimation Exploiting Covariate Independence in Two-Phase Randomized Trials
}
\description{
  This is the R package to calculate the semiparametric maximum
  likelihood estimate (SPMLE) and maximum estimated likelihood estimator
  (MELE) defined in Dai et al. (2009).
}
\details{
\tabular{ll}{
Package: \tab TwoPhaseInd\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2009-08-10\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab no\cr
}
Dai et al. (2009) proposed the algorithms of semiparametric maximum
likelihood estimate (SPMLE) and maximum estimated likelihood estimator
(MELE) for the treatment-biomarker interaction in two-phase sampling
nested within randomized clinical trials, incorporating the independence
between a randomized treatment and the baseline markers. This is the R
package providing the algorithms of SPMLE and MELE.
}
\author{
  James Y. Dai and Ting-Yuan Liu
}
\references{
J. Y. Dai, M. LeBlanc, and C. Kooperberg. Semiparametric estimation
exploiting co-variate independence in two-phase randomized
trials. Biometrics, 65(1):178-187, Mar 2009.
  
C. Kooperberg, M. Cushman, J. Hsia, J. G. Robinson, A. K. Aragaki,
J. K. Lynch, A. E. Baird, K. C. Johnson, L. H. Kuller, S. A. Beresford,
and B. Rodriguez. Can biomarkers identify women at increased stroke
risk? the women's health initiative hormone trials. PLoS clinical
trials, 2(6):e28, Jun 15 2007.   

}
\keyword{ package }
